+++ battle/entity/EEnemy.java
@@ -15,16 +15,16 @@ import java.util.ArrayList;
 
 public class EEnemy extends Entity {
 
-	public final int mark;
+	public final byte mark;
 	public final double mult, mula;
 
-	public int hit;
+	public byte hit;
 
 	public EEnemy(StageBasis b, MaskEnemy de, EAnimU ea, double magnif, double atkMagnif, int d0, int d1, int m) {
 		super(b, de, ea, atkMagnif, magnif);
 		mult = magnif;
 		mula = atkMagnif;
-		mark = m;
+		mark = (byte) m;
 		isBase = mark <= -1;
 		layer = d0 + (int) (b.r.nextDouble() * (d1 - d0 + 1));
 		traits = de.getTraits();
